requirements.txt
flask>=0.0
google-cloud-bigquery>=0.0
google-cloud-datastore>=0.0

